requirements.txt
openai==1.20.0
streamlit==1.33.0
python-dotenv==1.0.1
numpy
sounddevice
scipy 
